refactor: split states in files changes: - splited states in files for better maintenance - Added README for setting environment in another contributors - Created client to center some feats - Deleted unnecessary files, focus in Llama model
