+++ back-end/audio/text_to_speach.py
@@ -1,10 +1,12 @@
 from gtts import gTTS
 import os
+
 from config import constants
 
 class TextTSpeech():
     def __init__(self, text):
         self.text = text
+
 
     def convert_to_speech(self):
         tts = gTTS(text=self.text, lang='pt-br')
