+++ back-end/model/__init__.py
@@ -1,3 +1,1 @@
-from .gemini_analysis import GeminiThinking
-from .llama_analysis import LlamaThinking
 from .main_assistant import BusSenseAssistant
