+++ back-end/model/agents/__init__.py
@@ -1,3 +1,2 @@
 from .agent_feedback import FeedbackAgent
 from .agent_feedback_classifier import FeedbackClassifierAgent
-from .agent_phrase_generator import LlamaAgents
